+++ tests/agent/test_fargate_agent.py
@@ -32,6 +32,8 @@ def test_ecs_agent_config_options_default(monkeypatch, runner_token):
     assert not agent.subnets
     assert not agent.security_groups
     assert not agent.repository_credentials
+    assert not agent.task_role_arn
+    assert not agent.execution_role_arn
     assert agent.assign_public_ip == "ENABLED"
     assert agent.task_cpu == "256"
     assert agent.task_memory == "512"
@@ -47,6 +49,8 @@ def test_ecs_agent_config_options_init(monkeypatch, runner_token):
         aws_access_key_id="id",
         aws_secret_access_key="secret",
         aws_session_token="token",
+        task_role_arn="task_role_arn",
+        execution_role_arn="execution_role_arn",
         region_name="region",
         cluster="cluster",
         subnets=["subnet"],
@@ -58,6 +62,8 @@ def test_ecs_agent_config_options_init(monkeypatch, runner_token):
     )
     assert agent
     assert agent.name == "test"
+    assert agent.task_role_arn == "task_role_arn"
+    assert agent.execution_role_arn == "execution_role_arn"
     assert agent.cluster == "cluster"
     assert agent.subnets == ["subnet"]
     assert agent.security_groups == ["security_group"]
@@ -82,6 +88,8 @@ def test_ecs_agent_config_env_vars(monkeypatch, runner_token):
     monkeypatch.setenv("AWS_ACCESS_KEY_ID", "id")
     monkeypatch.setenv("AWS_SECRET_ACCESS_KEY", "secret")
     monkeypatch.setenv("AWS_SESSION_TOKEN", "token")
+    monkeypatch.setenv("TASK_ROLE_ARN", "task_role_arn")
+    monkeypatch.setenv("EXECUTION_ROLE_ARN", "execution_role_arn")
     monkeypatch.setenv("REGION_NAME", "region")
     monkeypatch.setenv("CLUSTER", "cluster")
     monkeypatch.setenv("REPOSITORY_CREDENTIALS", "repo")
@@ -91,6 +99,8 @@ def test_ecs_agent_config_env_vars(monkeypatch, runner_token):
 
     agent = FargateAgent(subnets=["subnet"])
     assert agent
+    assert agent.execution_role_arn == "execution_role_arn"
+    assert agent.task_role_arn == "task_role_arn"
     assert agent.cluster == "cluster"
     assert agent.repository_credentials == "repo"
     assert agent.assign_public_ip == "DISABLED"
@@ -164,6 +174,8 @@ def test_deploy_flows_all_args(monkeypatch, runner_token):
         aws_access_key_id="id",
         aws_secret_access_key="secret",
         aws_session_token="token",
+        task_role_arn="task_role_arn",
+        execution_role_arn="execution_role_arn",
         region_name="region",
         cluster="cluster",
         subnets=["subnet"],
@@ -300,6 +312,8 @@ def test_deploy_flows_register_task_definition_all_args(monkeypatch, runner_toke
         aws_access_key_id="id",
         aws_secret_access_key="secret",
         aws_session_token="token",
+        task_role_arn="task_role_arn",
+        execution_role_arn="execution_role_arn",
         region_name="region",
         cluster="cluster",
         subnets=["subnet"],